requirements.txt
fastapi
uvicorn
sentence-transformers
torch
faiss-cpu
scikit-learn
numpy
pandas
python-dotenv
